Rename 'database' package in iiss to 'reward_calc' and move reward calculation related modules
Rename 'icx_issue_formula' package to 'issue_formula'
