+++ tests/iiss/test_iiss_data_using_level_db.py
@@ -18,9 +18,9 @@ import plyvel
 import unittest
 from typing import TYPE_CHECKING
 
-from iconservice.iiss.msg_data import Header, GovernanceVariable, PRepsData, TxData, TxType, \
+from iconservice.iiss.reward_calc.msg_data import Header, GovernanceVariable, PRepsData, TxData, TxType, \
     DelegationTx, DelegationInfo, PRepRegisterTx, PRepUnregisterTx, BlockProduceInfoData
-from tests import create_address, rmtree
+from tests import create_address
 
 if TYPE_CHECKING:
     from iconservice.base.address import Address
