+++ tests/iiss/test_reward_calc_data_storage.py
@@ -18,21 +18,21 @@ import os
 import unittest
 from unittest.mock import patch
 
-from iconservice.iiss.data_creator import *
-from iconservice.iiss.msg_data import TxType
-from iconservice.iiss.reward_calc_data_storage import RewardCalcDataStorage
+from iconservice.iiss.reward_calc.data_creator import *
+from iconservice.iiss.reward_calc.msg_data import TxType
+from iconservice.iiss.reward_calc.data_storage import DataStorage as RewardCalcDatabase
 from tests import create_address
 from tests.iiss.mock_rc_db import MockIissDataBase
 from tests.mock_db import MockPlyvelDB
 
 
 class TestRcDataStorage(unittest.TestCase):
-    @patch('iconservice.iiss.database.iiss_db.IissDatabase.from_path')
+    @patch('iconservice.iiss.reward_calc.db.Database.from_path')
     @patch('os.path.exists')
     def setUp(self, _, mocked_iiss_db_from_path) -> None:
         self.path = ""
         mocked_iiss_db_from_path.side_effect = MockIissDataBase.from_path
-        self.rc_data_storage = RewardCalcDataStorage()
+        self.rc_data_storage = RewardCalcDatabase()
         self.rc_data_storage.open(self.path)
 
         dummy_block_height = 1
@@ -43,13 +43,13 @@ class TestRcDataStorage(unittest.TestCase):
 
         self.dummy_gv = GovernanceVariable()
         self.dummy_gv.block_height = dummy_block_height
-        self.dummy_gv.icx_price = 1
-        self.dummy_gv.incentive_rep = 10
+        self.dummy_gv.calculated_incentive_rep = 1
+        self.dummy_gv.reward_rep = 1000
 
         self.dummy_prep = PRepsData()
         self.dummy_prep.block_height = dummy_block_height
-        self.dummy_prep.block_generator = create_address()
-        self.dummy_prep.block_validator_list = [create_address()]
+        self.dummy_prep.total_delegation = 10
+        self.dummy_prep.prep_list = []
 
         self.dummy_tx = TxData()
         self.dummy_tx.address = create_address()
@@ -60,15 +60,15 @@ class TestRcDataStorage(unittest.TestCase):
     def tearDown(self):
         pass
 
-    @patch('iconservice.iiss.database.iiss_db.IissDatabase.from_path')
+    @patch('iconservice.iiss.reward_calc.db.Database.from_path')
     @patch('os.path.exists')
     def test_open(self, mocked_path_exists, mocked_iiss_db_from_path):
         # success case: when input existing path, make path of current_db and iiss_rc_db
         # and generate current level db(if not exist)
-        rc_data_storage = RewardCalcDataStorage()
+        rc_data_storage = RewardCalcDatabase()
         test_db_path: str = os.path.join(os.getcwd(), ".storage_test_db")
 
-        expected_current_db_path = os.path.join(test_db_path, RewardCalcDataStorage._CURRENT_IISS_DB_NAME)
+        expected_current_db_path = os.path.join(test_db_path, RewardCalcDatabase._CURRENT_IISS_DB_NAME)
 
         def from_path(path: str,
                       create_if_missing: bool = True) -> 'MockIissDataBase':
@@ -113,13 +113,13 @@ class TestRcDataStorage(unittest.TestCase):
         for block_height in range(-2, 1):
             self.assertRaises(AssertionError, self.rc_data_storage.create_db_for_calc, block_height)
 
-    @patch('iconservice.iiss.database.iiss_db.IissDatabase.from_path')
+    @patch('iconservice.iiss.reward_calc.db.Database.from_path')
     @patch('os.rename')
     @patch('os.path.exists')
     def test_create_db_for_calc_valid_block_height(self, mocked_path_exists, mocked_rename, mocked_iiss_db_from_path):
         # success case: when input valid block height, should create iiss_db and return path
         mocked_iiss_db_from_path.side_effect = MockIissDataBase.from_path
-        current_db_path = os.path.join(self.path, RewardCalcDataStorage._CURRENT_IISS_DB_NAME)
+        current_db_path = os.path.join(self.path, RewardCalcDatabase._CURRENT_IISS_DB_NAME)
 
         # todo: to be refactored
         def path_exists(path):
@@ -131,7 +131,7 @@ class TestRcDataStorage(unittest.TestCase):
 
         valid_block_height = 1
         expected_iiss_db_path = os.path.join(self.path,
-                                             RewardCalcDataStorage._IISS_RC_DB_NAME_PREFIX + f"{valid_block_height}")
+                                             RewardCalcDatabase._IISS_RC_DB_NAME_PREFIX + f"{valid_block_height}")
         actual_ret_path = self.rc_data_storage.create_db_for_calc(valid_block_height)
         self.assertEqual(expected_iiss_db_path, actual_ret_path)
