+++ tests/mock_generator.py
@@ -25,15 +25,15 @@ from iconservice.icon_config import default_icon_config
 from iconservice.icon_constant import ConfigKey
 from iconservice.icon_inner_service import IconScoreInnerTask
 from iconservice.icon_service_engine import IconServiceEngine
-from iconservice.iiss.database.db import Database
+from iconservice.iiss.reward_calc.db import Database as RewardCalcDatabase
 from tests import create_block_hash, rmtree
 
 SERVICE_ENGINE_PATH = 'iconservice.icon_service_engine.IconServiceEngine'
 ICX_ENGINE_PATH = 'iconservice.icx.icx_engine.IcxEngine'
 DB_FACTORY_PATH = 'iconservice.database.factory.ContextDatabaseFactory'
 ReqData = namedtuple("ReqData", "tx_hash, from_, to_, data_type, data")
-IISS_DB_PATH = 'iconservice.iiss.database.db'
-IISS_RC_DATA_STORAGE_PATH = 'iconservice.iiss.reward_calc_data_storage'
+IISS_DB_PATH = 'iconservice.iiss.reward_calc.db'
+IISS_RC_DATA_STORAGE_PATH = 'iconservice.iiss.reward_calc.data_storage'
 IISS_VARIABLE_PATH = 'iconservice.iiss.variable.variable'
 PREP_VARIABLE_PATH = 'iconservice.prep.variable.variable'
 
@@ -45,7 +45,7 @@ PREP_VARIABLE_PATH = 'iconservice.prep.variable.variable'
 @patch(f'{ICX_ENGINE_PATH}.open')
 @patch(f'{DB_FACTORY_PATH}.create_by_name')
 @patch(f'{IISS_DB_PATH}.Database.from_path')
-@patch(f'{IISS_RC_DATA_STORAGE_PATH}.RewardCalcDataStorage._load_last_transaction_index')
+@patch(f'{IISS_RC_DATA_STORAGE_PATH}.DataStorage._load_last_transaction_index')
 @patch(f'{IISS_VARIABLE_PATH}.Variable.init_config')
 @patch(f'{PREP_VARIABLE_PATH}.Variable.init_config')
 def generate_inner_task(
@@ -77,7 +77,7 @@ def generate_inner_task(
     context_db.get = state_get
     context_db.put = state_put
 
-    iiss_db = Mock(spec=Database)
+    iiss_db = Mock(spec=RewardCalcDatabase)
     iiss_db.get = iiss_get
     iiss_db.put = iiss_put
